+++ src/CfdiUtils/Elements/PagosAExtranjeros10/NoBeneficiario.php
@@ -1,0 +1,12 @@
+<?php
+namespace CfdiUtils\Elements\PagosAExtranjeros10;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class NoBeneficiario extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'pagosaextranjeros:NoBeneficiario';
+    }
+}